requirements.txt
flask==1.1.1
protobuf==3.11.3
mysql-connector-python==8.0.19